Update and rename Bullet to Bullet.java

+++ GUI/Component/Bullet.java
@@ -5,7 +5,7 @@ import java.awt.geom.Point2D;
 public abstract class Bullet extends GameObject
 {
 protected final int myDamage;
-protected int mySpeed;
+protected double mySpeed;
 protected Player myPlayer;
 protected Long startTime;
